// apps/api/src/workflows/dto/save-workflow.dto.ts
export class SaveNodeDto {
  id: string;
  templateId: string;
  label?: string;
  positionX: number;
  positionY: number;
  actionState: Record<string, any>;
}

export class SaveEdgeDto {
  id: string;
  sourceNodeId: string;
  targetNodeId: string;
  sourceHandle?: string;
  targetHandle?: string;
}

export class UpdateWorkflowDto {
  name?: string;
  description?: string;
  status?: string;
  nodes: SaveNodeDto[];
  edges: SaveEdgeDto[];
}